Fix the issue of setting the height of comment details to 0 incorrectly (Win Linux Switch) This issue is due to SIZE_T_MAX was set to an incorrect value on a 64 bit system. This fix adds judgment for 64 bit system and ensures that even if the SIZE_T_MAX is set to an incorrect value, it will not cause similar problems again

+++ wiliwili/source/activity/player_activity.cpp
@@ -387,7 +387,7 @@ void PlayerActivity::onUGCSeasonInfo(const bilibili::UGCSeason& result) {
                 auto* item =
                     (PlayerTabCell*)recycler->dequeueReusableCell("Cell");
                 item->title->setText(d.title);
-                item->setSelected(ds->getCurrentIndex() == d.index);
+                item->setSelected(ds->getCurrentIndex() == (size_t)d.index);
                 item->setBadge(
                     wiliwili::sec2MinSec(d.page.duration), nvgRGBA(0, 0, 0, 0),
                     brls::Application::getTheme().getColor("font/grey"));
